[WB-1973-switch] Colocate base styles
